+++ packages/sdk/CHANGELOG.md
@@ -13,6 +13,11 @@ commit that affects the code in a way that consumers might care about should inc
 
 -   Update documentation for `FieldType.CHECKBOX` cell read & write types to be more accurate.
 -   Fetching field types & configs is now cached, making it more performant.
+-   Field descriptions can now be edited and specified when creating a new field:
+    -   Added `field.updateDescriptionAsync`.
+    -   Added optional `description` argument to `table.createFieldAsync`.
+    -   Added optional `description` property to `field` objects accepted by
+        `base.createTableAsync`.
 
 ## [1.8.0](https://github.com/airtable/blocks/compare/@airtable/blocks@1.7.2...@airtable/blocks@1.8.0) - 2021-07-12